curso2/pyproject.toml
[project]
name = "curso2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "numpy>=2.3.0",
    "openai>=1.84.0",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
    "uuid>=1.30",
]
